feat: remove login & update user schema
