+++ app/(protected)/user-profile/Guide.tsx
