+++ components/Header.tsx
@@ -28,7 +28,7 @@ export default function Header() {
           animate={{ opacity: 1, x: 0 }}
           transition={{ delay: 0.2, duration: 0.5 }}
         >
-          <SignIn />
+          {/* <SignIn /> */}
         </motion.div>
       </div>
     </motion.div>
