+++ types/auth.d.ts
@@ -7,6 +7,7 @@ declare module "next-auth" {
       name: string | null;
       email: string | null;
       image: string | null;
+      total_commits: number | null;
     };
   }
